requirements.txt
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
pandas==2.1.3
pylatex==1.4.1
matplotlib==3.8.2
seaborn==0.13.0
biopython